good riddens to tuple_cmp
now unused.

This function is too depressing to read to the end.

The entire body could have been something like

    return cmp(tuple(map(int,t1)), tuple(map(int,t2)))

rather than a screenful of pain.

Nevermind why it found its way specifically into TestStartGratia and is
implemented as an instance method.

Anyway, don't need it anymore, so it's going away.